trustaid-backend-win-v4d/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
chromadb==0.5.3
openai==1.35.7
sqlglot>=25
python-dotenv==1.0.1
httpx==0.27.2